增加通讯

* LHB-通讯修改: (9 commits squashed)

  - 通讯模块代码完善

  - 完善通讯

  - Merge branch &#39;Develop&#39; into LHB-通讯修改

  - 完善通讯模块，增加UDP\TCP插件

  - 通讯模块完善

  - 1

  - 通讯修改

  - 1

  - 合并修改

Merged-by: 汤传刚 <tangcg@foxmail.com>

CR-link: https://codeup.aliyun.com/6479fd8ed2963c5649be0e69/SoftWareDevelopers/Metal_Burr_Detection/Metal_Burr_TCG/change/34

+++ MarkControl/CMarkCtrlVM.cs
@@ -130,19 +130,19 @@ namespace MarkControl
             CMiniEcatLib.Mb_E4O4Encoder_Initial(
                 MarkConfig.SlaveId,
                 MarkConfig.EncoderId,
-                MarkConfig.EncoderMode,
-                MarkConfig.EncoderDir,
+                (int)MarkConfig.EncoderMode,
+                (int)MarkConfig.EncoderDir,
                 1
             );
             //编码器数值清零,打开软件时跟相机同步清零
             CMiniEcatLib.Mb_E4O4Encoder_SetCurrentData(MarkConfig.SlaveId, MarkConfig.EncoderId, 0);
             //设置触发输出模式
-            CMiniEcatLib.Mb_E4O4TrigOut_SetOutMode(MarkConfig.SlaveId, MarkConfig.OutMode);
+            CMiniEcatLib.Mb_E4O4TrigOut_SetOutMode(MarkConfig.SlaveId, (int)MarkConfig.OutMode);
             //设置触发模式
             CMiniEcatLib.Mb_E4O4TrigOut_SetTrigMode(
                 MarkConfig.SlaveId,
                 MarkConfig.TriggerId,
-                MarkConfig.TrigMode
+                (int)MarkConfig.TrigMode
             );
             //设置脉宽
             CMiniEcatLib.Mb_E4O4TrigOut_SetPulseWidth(
